update 9 files, delete 1 file, rename 2 files and create 1 file

+++ src/app/_types/CartItem.ts
@@ -1,8 +1,10 @@
-import { z } from "zod";
+import { z } from "zod"; // zodライブラリのインポート
 
+// バリデーションスキーマ（データが満たすべき制約条件や構造を定義）
 export const cartItemSchema = z.object({
-  productId: z.string().min(1), // 1文字以上
-  quantity: z.number().min(0), // 0以上
+  productId: z.string().min(1), // 1文字以上の「文字列」
+  quantity: z.number().int().min(0), // 0以上の「整数値」
 });
 
+// バリデーションスキーマをもとに「CartItem型」を生成
 export type CartItem = z.infer<typeof cartItemSchema>;
